Moved the SQL execute statments  to a new shared floder. Changed design of my clsSQL and update methods to work with new class
