+++ FootlockerInvoiceApp/FootlockerInvoiceApp.csproj
@@ -60,6 +60,7 @@
     <Compile Include="Search\wndSearch.xaml.cs">
       <DependentUpon>wndSearch.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Shared\clsDatabase.cs" />
     <Page Include="Items\wndItems.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
